Move exception handler tests to a new test suite
