+++ libcaf_core/caf/scheduled_actor.test.cpp
@@ -6,6 +6,11 @@
 
 #include "caf/test/test.hpp"
 
+#include "caf/actor_system.hpp"
+#include "caf/actor_system_config.hpp"
+#include "caf/event_based_actor.hpp"
+#include "caf/scoped_actor.hpp"
+
 using namespace caf;
 
 #define ASSERT_COMPILES(expr, msg)                                             \
@@ -135,6 +140,63 @@ struct const_exception_fn {
 
 ASSERT_COMPILES(set_exception_handler(const_exception_fn{}),
                 "set_exception_handler must accept const function objects");
+
+behavior testee() {
+  return {
+    [](const std::string&) { throw std::runtime_error("whatever"); },
+  };
+}
+
+behavior exception_testee(event_based_actor* self) {
+  self->set_exception_handler([](std::exception_ptr&) -> error {
+    return exit_reason::remote_link_unreachable;
+  });
+  return testee();
+}
+
+TEST("the default exception handler includes the error message") {
+  using std::string;
+  actor_system_config cfg;
+  actor_system system{cfg};
+  scoped_actor self{system};
+  auto aut = self->spawn(testee);
+  self->mail("hello world")
+    .request(aut, infinite)
+    .receive( //
+      [this] { fail("unexpected response"); },
+      [this](const error& err) {
+        check_eq(err.what(),
+                 "unhandled exception of type std.runtime_error: whatever");
+      });
+}
+
+TEST("actors can override the default exception handler") {
+  actor_system_config cfg;
+  actor_system system{cfg};
+  auto handler = [](std::exception_ptr& eptr) -> error {
+    try {
+      std::rethrow_exception(eptr);
+    } catch (std::runtime_error&) {
+      return exit_reason::normal;
+    } catch (...) {
+      // "fall through"
+    }
+    return sec::runtime_error;
+  };
+  scoped_actor self{system};
+  auto testee1 = self->spawn<monitored>([=](event_based_actor* eb_self) {
+    eb_self->set_exception_handler(handler);
+    throw std::runtime_error("ping");
+  });
+  auto testee2 = self->spawn<monitored>([=](event_based_actor* eb_self) {
+    eb_self->set_exception_handler(handler);
+    throw std::logic_error("pong");
+  });
+  auto testee3 = self->spawn<monitored>(exception_testee);
+  self->mail("foo").send(testee3);
+  // receive all down messages
+  self->wait_for(testee1, testee2, testee3);
+}
 
 #endif // CAF_ENABLE_EXCEPTIONS
